% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_from_files.R
\name{stream_from_files}
\alias{stream_from_files}
\title{Generates a two dimensional data stream from data files in a given folder.}
\usage{
stream_from_files(folder)
}
\arguments{
\item{folder}{The folder with the data files.}
}
\description{
Generates a two dimensional data stream from data files in a given folder.
}
\examples{
folder <- tempdir()
out <- gen_stream(2, folder = folder)
stream <- stream_from_files(paste(folder, "/data", sep=""))
dim(stream)
unlink(folder, recursive = TRUE)
}
\seealso{
\code{\link{gen_stream}}.
}
